Move workspace form inside workspace creator
