+++ src/plugins/workspace/public/components/workspace_form/fields/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { WorkspaceDescriptionField } from './workspace_description_field';
+export { WorkspaceNameField } from './workspace_name_field';
