+++ src/plugins/workspace/public/components/workspace_form/index.ts
@@ -3,12 +3,31 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { WorkspaceForm } from './workspace_form';
 export { WorkspaceDetailForm } from './workspace_detail_form';
-export { WorkspaceFormSubmitData } from './types';
-export { WorkspaceOperationType } from './constants';
+export { SelectDataSourcePanel } from './select_data_source_panel';
+export { WorkspaceFormErrorCallout } from './workspace_form_error_callout';
+export { WorkspaceUseCase } from './workspace_use_case';
+export { WorkspacePermissionSettingPanel } from './workspace_permission_setting_panel';
+export { WorkspaceCancelModal } from './workspace_cancel_modal';
+export { WorkspaceNameField, WorkspaceDescriptionField } from './fields';
+export { EnterDetailsPanel } from './workspace_enter_details_panel';
+
+export { WorkspaceFormSubmitData, WorkspaceFormData, WorkspaceFormProps } from './types';
+export {
+  WorkspaceOperationType,
+  DetailTab,
+  DetailTabTitles,
+  WorkspacePermissionItemType,
+  usersAndPermissionsCreatePageTitle,
+  selectDataSourceTitle,
+  workspaceDetailsTitle,
+  workspaceUseCaseTitle,
+} from './constants';
 export {
   convertPermissionsToPermissionSettings,
   convertPermissionSettingsToPermissions,
 } from './utils';
+
 export { WorkspaceFormProvider, useWorkspaceFormContext } from './workspace_form_context';
+export { useWorkspaceForm } from './use_workspace_form';
+export { useFormAvailableUseCases } from './use_form_available_use_cases';
